requirements.txt
tqdm
aspose-cad